Add support for metadata in figures without captions
Enhances the handling of figures that do not have user-provided captions by automatically adding an unnumbered caption node, allowing metadata (such as license or attribution) to be displayed in the caption area. Updates the documentation with new examples and refines the logic for appending metadata to captions. Also connects a new event handler to ensure unnumbered captions are added during doctree resolution.